fix: 번들 생성 시 createdAt 비는 문제 해결, ShareGroupScheduler 분리
