+++ src/main/java/ssu/today/domain/diary/scheduler/DiaryBundleScheduler.java
@@ -117,4 +117,3 @@ public class DiaryBundleScheduler {
 
 
 }
-
